#73 - correções das configurações de segurança
