--- src/main/java/br/com/academiadev/bumblebee/config/SecurityConfig.java
@@ -1,107 +1,0 @@
-package br.com.academiadev.bumblebee.config;
-
-import br.com.academiadev.bumblebee.model.Usuario;
-import br.com.academiadev.bumblebee.repository.UsuarioRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableAuthorizationServer;
-import org.springframework.security.oauth2.provider.ClientDetailsService;
-import org.springframework.security.oauth2.provider.approval.ApprovalStore;
-import org.springframework.security.oauth2.provider.approval.TokenApprovalStore;
-import org.springframework.security.oauth2.provider.approval.TokenStoreUserApprovalHandler;
-import org.springframework.security.oauth2.provider.request.DefaultOAuth2RequestFactory;
-import org.springframework.security.oauth2.provider.token.TokenStore;
-import org.springframework.security.oauth2.provider.token.store.InMemoryTokenStore;
-
-@Configuration
-@EnableAuthorizationServer
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Autowired
-    private ClientDetailsService clientDetailsService;
-
-    @Bean
-    private static BCryptPasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Autowired
-    protected void configure(AuthenticationManagerBuilder auth, UsuarioRepository repository) throws Exception {
-        Usuario usuario = new Usuario();
-        usuario.setSenha("adminadmin");
-        usuario.setEmail("admin@admin.com");
-        usuario.setNome("Administrador do sistema");
-        usuario.setEnable(true);
-
-        if (repository.count() == 0)
-            repository.saveAndFlush(usuario);
-
-        auth.userDetailsService(email -> repository.findByEmail(email).orElse(null));
-    }
-
-    @Override
-    public void configure(WebSecurity web) throws Exception {
-        web.ignoring().antMatchers(HttpMethod.GET, //
-                "/", //
-                "/webjars/**", //
-                "/*.html", //
-                "/favicon.ico", //
-                "/**/*.html", //
-                "/v2/api-docs", //
-                "/configuration/ui", //
-                "/swagger-resources/**", //
-                "/configuration/**", //
-                "/swagger-ui.html", //
-                "/webjars/**", //
-                "/**/*.css", //
-                "/**/*.js"//
-        );
-
-    }
-
-//    @Override
-//    protected void configure(HttpSecurity http) throws Exception {
-//        http.csrf()
-//                .disable()
-//                .anonymous().disable()
-//                .authorizeRequests()
-//                .antMatchers("/oauth/token").permitAll();
-//    }
-
-    @Override
-    @Bean
-    public AuthenticationManager authenticationManagerBean() throws Exception {
-        return super.authenticationManagerBean();
-    }
-
-
-    @Bean
-    public TokenStore tokenStore() {
-        return new InMemoryTokenStore();
-    }
-
-    @Bean
-    @Autowired
-    public TokenStoreUserApprovalHandler userApprovalHandler(TokenStore tokenStore) {
-        TokenStoreUserApprovalHandler handler = new TokenStoreUserApprovalHandler();
-        handler.setTokenStore(tokenStore);
-        handler.setRequestFactory(new DefaultOAuth2RequestFactory(clientDetailsService));
-        handler.setClientDetailsService(clientDetailsService);
-        return handler;
-    }
-
-    @Bean
-    @Autowired
-    public ApprovalStore approvalStore(TokenStore tokenStore) throws Exception {
-        TokenApprovalStore store = new TokenApprovalStore();
-        store.setTokenStore(tokenStore);
-        return store;
-    }
-}
